+++ lib/src/model/snacky.dart
@@ -18,6 +18,7 @@ class Snacky {
   final Duration transitionDuration;
   final Curve transitionCurve;
   final SnackyLocation? location;
+  final EdgeInsets? padding;
 
   final Widget Function(BuildContext, CancelableSnacky)? builder;
   final EdgeInsetsGeometry? margin;
@@ -43,6 +44,7 @@ class Snacky {
     this.transitionCurve = Curves.easeInOut,
     this.margin,
     this.location,
+    this.padding,
   })  : builder = null,
         _showDuration = showDuration;
 
@@ -53,6 +55,9 @@ class Snacky {
     this.transitionCurve = Curves.easeInOut,
     this.margin,
     this.location,
+    this.openUntillClosed = false,
+    this.canBeClosed = false,
+    this.padding,
   })  : title = '',
         _showDuration = showDuration,
         type = SnackyType.info,
@@ -60,7 +65,5 @@ class Snacky {
         leadingWidgetBuilder = null,
         trailingWidgetBuilder = null,
         bottomWidgetBuilder = null,
-        onTap = null,
-        canBeClosed = false,
-        openUntillClosed = false;
+        onTap = null;
 }