+++ lib/src/widget/snacky_configurator_widget.dart
@@ -55,17 +55,20 @@ class _SnackyConfiguratorWidgetState extends State<SnackyConfiguratorWidget>
     final snackyLocation =
         layoutConfig.getSnackyLocation(context, activeSnacky.snacky);
     return Builder(
-      builder: (context) => Stack(
-        key: ValueKey(activeSnacky.hashCode),
-        alignment: snackyLocation.alignment,
-        children: [
-          widget.snackyBuilder.build(
-            context,
-            layoutConfig,
-            activeSnacky,
-            snackyController,
-          ),
-        ],
+      builder: (context) => Padding(
+        padding: activeSnacky.snacky.padding ?? EdgeInsets.zero,
+        child: Stack(
+          key: ValueKey(activeSnacky.hashCode),
+          alignment: snackyLocation.alignment,
+          children: [
+            widget.snackyBuilder.build(
+              context,
+              layoutConfig,
+              activeSnacky,
+              snackyController,
+            ),
+          ],
+        ),
       ),
     );
   }