+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.8.0
+
+## Feat
+
+- Added `padding` to `Snacky`
+- Added `openUntilClosed` and `canBeClosed` as an option for a `Snacky.widget`
+
 # 0.7.0
 
 ## Feat